requirements.txt
APScheduler==3.9.1
Django==4.1.1
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
django-environ==0.9.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
pymongo==4.2.0
redis==4.3.4
requests==2.28.1
Scrapy==2.6.2
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1

itemadapter~=0.7.0
